feat: create new sdk for v2
